Put sort function in firebase watch function to sort allTrees by distance
